Add Software TPM
The Software TPM has two components - the TPM2 table
and device nodes. The TPM2 table advertises a TPM2 module
is present. We add the firmware TPM emulator
device node which will serve as the TPM backend.

Note: Software TPM is emulation and, as such, is not
fit for production security use.